requirements.txt
insightface
onnxruntime
opencv-python
numpy
Pillow
customtkinter
sounddevice
